Revert "Merge pull request #9 from akashpatil7/G81-backend-setup"
Reverting this because it was undo-ing some previous changes that made the user services work with the frontend. It no longer works if you include this commit. We can discuss it further in the meeting tomorrow if this is an issue.

This reverts commit dac2630318fa623c47fa529e4c1b4c98796beafe, reversing
changes made to 95d01033167f88ddb579fadbdf8387944b2f233f.